strands/requirements.txt
strands-agents
bedrock-agentcore
requests
aws-opentelemetry-distro>=0.10.0